Changed the category system.
Now the categories use generic codes and only accept enumerations. This
reduces the code required by the user of the library.

Updated code accordingly.

Added more tests for the logs.

+++ projects/Sdl2Sharp.Examples/Sdl2Sharp.Examples.Example001.HelloSdl/src/Program.cs
@@ -2,9 +2,9 @@
 using Sdl2Sharp.Utilities;
 using Sdl2Sharp.Utilities.Log;
 
-Functions.SetLogPriority((int)MyCategories.LibraryInformation, Priority.Information);
-Functions.SetLogPriority((int)MyCategories.SubsystemInformation, Priority.Information);
-Functions.SetLogPriority((int)MyCategories.CpuInformation, Priority.Information);
+Functions.SetLogPriority(MyCategories.LibraryInformation, Priority.Information);
+Functions.SetLogPriority(MyCategories.SubsystemInformation, Priority.Information);
+Functions.SetLogPriority(MyCategories.CpuInformation, Priority.Information);
 
 Functions.SetLogOutputFunction((data, category, priority, message) =>
 {
@@ -41,30 +41,30 @@ Functions.SetLogOutputFunction((data, category, priority, message) =>
     Console.ResetColor();
 }, DateTime.Now);
 
-Functions.LogInformation((int)MyCategories.LibraryInformation,
+Functions.LogInformation(MyCategories.LibraryInformation,
     $"Using SDL v{NativeDllInformation.Version} [{NativeDllInformation.Revision}] on {Platform.Name}");
 
-Functions.LogInformation((int)MyCategories.CpuInformation, $"Logical Cores: {CpuInformation.LogicalCores}");
-Functions.LogInformation((int)MyCategories.CpuInformation, $"Cache Line Size: {CpuInformation.CacheLineSize}");
-Functions.LogInformation((int)MyCategories.CpuInformation, $"System RAM: {CpuInformation.SystemRam}MB");
-Functions.LogInformation((int)MyCategories.CpuInformation, $"CPU has the RDTSC instruction: {CpuInformation.HasRdtsc}");
-Functions.LogInformation((int)MyCategories.CpuInformation, $"CPU has AltiVec features: {CpuInformation.HasAltiVec}");
-Functions.LogInformation((int)MyCategories.CpuInformation, $"CPU has MMX features: {CpuInformation.HasMmx}");
-Functions.LogInformation((int)MyCategories.CpuInformation, $"CPU has 3DNow! features: {CpuInformation.Has3DNow}");
-Functions.LogInformation((int)MyCategories.CpuInformation, $"CPU has SSE features: {CpuInformation.HasSse}");
-Functions.LogInformation((int)MyCategories.CpuInformation, $"CPU has SSE2 features: {CpuInformation.HasSse2}");
-Functions.LogInformation((int)MyCategories.CpuInformation, $"CPU has SSE3 features: {CpuInformation.HasSse3}");
-Functions.LogInformation((int)MyCategories.CpuInformation, $"CPU has SSE4.1 features: {CpuInformation.HasSse41}");
-Functions.LogInformation((int)MyCategories.CpuInformation, $"CPU has SSE4.2 features: {CpuInformation.HasSse42}");
-Functions.LogInformation((int)MyCategories.CpuInformation, $"CPU has AVX features: {CpuInformation.HasAvx}");
-Functions.LogInformation((int)MyCategories.CpuInformation, $"CPU has AVX2 features: {CpuInformation.HasAvx2}");
-Functions.LogInformation((int)MyCategories.CpuInformation, $"CPU has AVX-512F features: {CpuInformation.HasAvx512F}");
-Functions.LogInformation((int)MyCategories.CpuInformation, $"CPU has ARM SIMD features: {CpuInformation.HasArmSimd}");
-Functions.LogInformation((int)MyCategories.CpuInformation, $"CPU has NEON features: {CpuInformation.HasNeon}");
+Functions.LogInformation(MyCategories.CpuInformation, $"Logical Cores: {CpuInformation.LogicalCores}");
+Functions.LogInformation(MyCategories.CpuInformation, $"Cache Line Size: {CpuInformation.CacheLineSize}");
+Functions.LogInformation(MyCategories.CpuInformation, $"System RAM: {CpuInformation.SystemRam}MB");
+Functions.LogInformation(MyCategories.CpuInformation, $"CPU has the RDTSC instruction: {CpuInformation.HasRdtsc}");
+Functions.LogInformation(MyCategories.CpuInformation, $"CPU has AltiVec features: {CpuInformation.HasAltiVec}");
+Functions.LogInformation(MyCategories.CpuInformation, $"CPU has MMX features: {CpuInformation.HasMmx}");
+Functions.LogInformation(MyCategories.CpuInformation, $"CPU has 3DNow! features: {CpuInformation.Has3DNow}");
+Functions.LogInformation(MyCategories.CpuInformation, $"CPU has SSE features: {CpuInformation.HasSse}");
+Functions.LogInformation(MyCategories.CpuInformation, $"CPU has SSE2 features: {CpuInformation.HasSse2}");
+Functions.LogInformation(MyCategories.CpuInformation, $"CPU has SSE3 features: {CpuInformation.HasSse3}");
+Functions.LogInformation(MyCategories.CpuInformation, $"CPU has SSE4.1 features: {CpuInformation.HasSse41}");
+Functions.LogInformation(MyCategories.CpuInformation, $"CPU has SSE4.2 features: {CpuInformation.HasSse42}");
+Functions.LogInformation(MyCategories.CpuInformation, $"CPU has AVX features: {CpuInformation.HasAvx}");
+Functions.LogInformation(MyCategories.CpuInformation, $"CPU has AVX2 features: {CpuInformation.HasAvx2}");
+Functions.LogInformation(MyCategories.CpuInformation, $"CPU has AVX-512F features: {CpuInformation.HasAvx512F}");
+Functions.LogInformation(MyCategories.CpuInformation, $"CPU has ARM SIMD features: {CpuInformation.HasArmSimd}");
+Functions.LogInformation(MyCategories.CpuInformation, $"CPU has NEON features: {CpuInformation.HasNeon}");
 
 using (new Subsystem(SubsystemFlags.Video))
 {
-    Functions.LogInformation((int)MyCategories.SubsystemInformation,
+    Functions.LogInformation(MyCategories.SubsystemInformation,
         $"Running SDL with subsystems [{Subsystem.InitializedSubsystems}]");
 }
